+++ extensions/prompt-stash/CHANGELOG.md
@@ -1,5 +1,9 @@
 # PromptStash Changelog
 
+## [Feature] - {PR_MERGE_DATE}
+
+- Added a preference to automatically insert the selected prompt into the text field when creating a new prompt.
+
 ## [Feature] - 2025-01-07
 
 - Added an edit prompt action to the prompt list.